# client/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    url: https://github.com/moov-io/watchman
  description: Moov Watchman is an HTTP API and Go library to download, parse and
    offer search functions over numerous trade sanction lists from the United States,
    European Union governments, agencies, and non profits for complying with regional
    laws. Also included is a web UI and async webhook notification service to initiate
    processes on remote systems.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Watchman API
  version: v1
servers:
- description: Local development
  url: http://localhost:8084
...
paths:
...
  /ofac/sdn/{sdnID}:
    get:
      description: Get SDN details
      operationId: getSDN
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: SDN ID
        explode: false
        in: path
        name: sdnID
        required: true
        schema:
          example: 564dd7d1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacSDN'
          description: SDN metadata
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Get SDN
      tags:
      - Watchman
  /ofac/sdn/{sdnID}/alts:
    get:
      operationId: getSDNAltNames
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: SDN ID
        explode: false
        in: path
        name: sdnID
        required: true
        schema:
          example: 564dd7d1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacSDNAltNames'
          description: SDN alternate names
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Get SDN alt names
      tags:
      - Watchman
  /ofac/sdn/{sdnID}/addresses:
    get:
      operationId: getSDNAddresses
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: SDN ID
        explode: false
        in: path
        name: sdnID
        required: true
        schema:
          example: 564dd7d1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacEntityAddresses'
          description: SDN addresses
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Get SDN addresses
      tags:
      - Watchman
...
  /downloads:
    get:
      description: Return list of recent downloads of list data
      operationId: getLatestDownloads
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Maximum number of downloads to return sorted by their timestamp
          in decending order.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          example: 25
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Downloads'
          description: Recent timestamps and counts of parsed objects
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Get latest downloads
      tags:
      - Watchman
...
components:
  schemas:
...
    OfacSDN:
      description: Specially designated national from OFAC list
      example:
        match: 0.91
        entityID: "1231"
        programs:
        - CUBA
        sdnName: BANCO NACIONAL DE CUBA
        title: Title of an individual
        remarks: remarks
      properties:
        entityID:
          example: "1231"
          type: string
        sdnName:
          example: BANCO NACIONAL DE CUBA
          type: string
        sdnType:
          $ref: '#/components/schemas/SdnType'
        programs:
          description: Programs is the sanction programs this SDN was added from
          example:
          - CUBA
          items:
            type: string
          type: array
        title:
          example: Title of an individual
          type: string
        remarks:
          type: string
        match:
          description: Remarks on SDN and often additional information about the SDN
          example: 0.91
          type: number
    OfacEntityAddresses:
      items:
        $ref: '#/components/schemas/OfacEntityAddress'
      type: array
    OfacEntityAddress:
      description: Physical address from OFAC list
      example:
        country: Japan
        address: 123 73th St
        cityStateProvincePostalCode: Tokyo 103
        match: 0.91
        entityID: "2112"
        addressID: "201"
      properties:
        entityID:
          example: "2112"
          type: string
        addressID:
          example: "201"
          type: string
        address:
          example: 123 73th St
          type: string
        cityStateProvincePostalCode:
          example: Tokyo 103
          type: string
        country:
          example: Japan
          type: string
        match:
          example: 0.91
          type: number
    OfacSDNAltNames:
      items:
        $ref: '#/components/schemas/OfacAlt'
      type: array
    OfacAlt:
      description: Alternate name from OFAC list
      example:
        alternateID: "220"
        alternateRemarks: Extra information
        alternateType: aka
        match: 0.91
        entityID: "306"
        alternateName: NATIONAL BANK OF CUBA
      properties:
        entityID:
          example: "306"
          type: string
        alternateID:
          example: "220"
          type: string
        alternateType:
          example: aka
          type: string
        alternateName:
          example: NATIONAL BANK OF CUBA
          type: string
        alternateRemarks:
          example: Extra information
          type: string
        match:
          example: 0.91
          type: number
    SdnType:
      description: Used for classifying SDNs — typically represents an individual
        or company
      enum:
      - individual
      - entity
      - vessel
      - aircraft
      type: string
...
    Downloads:
      items:
        $ref: '#/components/schemas/Download'
      type: array
    Download:
      description: Metadata and stats about downloaded OFAC data
      example:
        SDNs: 7414
        altNames: 9729
        addresses: 11747
        deniedPersons: 842
        bisEntities: 1391
        sectoralSanctions: 329
        timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        SDNs:
          example: 7414
          type: integer
        altNames:
          example: 9729
          type: integer
        addresses:
          example: 11747
          type: integer
        sectoralSanctions:
          example: 329
          type: integer
        deniedPersons:
          example: 842
          type: integer
        bisEntities:
          example: 1391
          type: integer
        timestamp:
          format: date-time
          type: string
...
    Error:
      properties:
        error:
          description: An error message describing the problem intended for humans.
          example: Example error, see description
          type: string
      required:
      - error